requirements.txt
Flask==1.0.2
jsonschema==2.6.0
flask-caching==1.3.3
marshmallow==3.0.0b13
flask-jwt==0.3.2
flask-sqlalchemy==2.3.1
strict-rfc3339==0.7
psycopg2==2.7.5
python-dateutil==2.7.3